new class: attributes
